ML/cloud/requirement.txt
google-cloud-storage
google-cloud-speech
functions-framework
moviepy
pydub
pandas
scikit-learn
numpy
